refactor: policies to envVersion
